requirements.txt
aiogram==2.23.1
aioschedule
apiclient
asyncio
beautifulsoup4
python-dotenv
httplib2
logging
lxml
numpy
oauth2client
openpyxl
pandas
requests
selenium
users
google-api-python-client
webdriver_manager
